move openapi modules into subfolder
