+++ src/classifications/knn.journey.test.ts
@@ -1,6 +1,6 @@
 import weaviate, { WeaviateClient } from '../index';
 import Connection from '../connection';
-import { Classification } from '../types';
+import { Classification } from '../openapi/types';
 
 const targetDessertId = 'cd54852a-209d-423b-bf1c-884468215237';
 const targetSavoryId = 'e5da0127-327e-4184-85b8-7b9d1af4a850';
