+++ src/openapi/schema.ts
